fmincon running
The function is now working.
Or at least it is printing values :)

The contour plot is still printing the same from the example with the fuel optimization. that has to be fixed.

We need to implement constraints for example osat. - more can be optimized.

I leave i for this morning. have an assignment in another course.